+++ libsolidity/analysis/ContractLevelChecker.cpp
@@ -41,8 +41,6 @@ bool ContractLevelChecker::check(ContractDefinition const& _contract)
 	checkIllegalOverrides(_contract);
 	checkAbstractFunctions(_contract);
 	checkBaseConstructorArguments(_contract);
-	checkConstructor(_contract);
-	checkFallbackFunction(_contract);
 	checkExternalTypeClashes(_contract);
 	checkHashCollisions(_contract);
 	checkLibraryRequirements(_contract);
@@ -58,6 +56,7 @@ void ContractLevelChecker::checkDuplicateFunctions(ContractDefinition const& _co
 	map<string, vector<FunctionDefinition const*>> functions;
 	FunctionDefinition const* constructor = nullptr;
 	FunctionDefinition const* fallback = nullptr;
+	FunctionDefinition const* etherReceiver = nullptr;
 	for (FunctionDefinition const* function: _contract.definedFunctions())
 		if (function->isConstructor())
 		{
@@ -78,6 +77,16 @@ void ContractLevelChecker::checkDuplicateFunctions(ContractDefinition const& _co
 					"Only one fallback function is allowed."
 				);
 			fallback = function;
+		}
+		else if (function->isEtherReceiver())
+		{
+			if (etherReceiver)
+				m_errorReporter.declarationError(
+					function->location(),
+					SecondarySourceLocation().append("Another declaration is here:", etherReceiver->location()),
+					"Only one ether receiver function is allowed."
+				);
+			etherReceiver = function;
 		}
 		else
 		{
@@ -356,48 +365,6 @@ void ContractLevelChecker::annotateBaseConstructorArguments(
 		);
 	}
 
-}
-
-void ContractLevelChecker::checkConstructor(ContractDefinition const& _contract)
-{
-	FunctionDefinition const* constructor = _contract.constructor();
-	if (!constructor)
-		return;
-
-	if (!constructor->returnParameters().empty())
-		m_errorReporter.typeError(constructor->returnParameterList()->location(), "Non-empty \"returns\" directive for constructor.");
-	if (constructor->stateMutability() != StateMutability::NonPayable && constructor->stateMutability() != StateMutability::Payable)
-		m_errorReporter.typeError(
-			constructor->location(),
-			"Constructor must be payable or non-payable, but is \"" +
-			stateMutabilityToString(constructor->stateMutability()) +
-			"\"."
-		);
-	if (constructor->visibility() != FunctionDefinition::Visibility::Public && constructor->visibility() != FunctionDefinition::Visibility::Internal)
-		m_errorReporter.typeError(constructor->location(), "Constructor must be public or internal.");
-}
-
-void ContractLevelChecker::checkFallbackFunction(ContractDefinition const& _contract)
-{
-	FunctionDefinition const* fallback = _contract.fallbackFunction();
-	if (!fallback)
-		return;
-
-	if (_contract.isLibrary())
-		m_errorReporter.typeError(fallback->location(), "Libraries cannot have fallback functions.");
-	if (fallback->stateMutability() != StateMutability::NonPayable && fallback->stateMutability() != StateMutability::Payable)
-		m_errorReporter.typeError(
-			fallback->location(),
-			"Fallback function must be payable or non-payable, but is \"" +
-			stateMutabilityToString(fallback->stateMutability()) +
-			"\"."
-	);
-	if (!fallback->parameters().empty())
-		m_errorReporter.typeError(fallback->parameterList().location(), "Fallback function cannot take parameters.");
-	if (!fallback->returnParameters().empty())
-		m_errorReporter.typeError(fallback->returnParameterList()->location(), "Fallback function cannot return values.");
-	if (fallback->visibility() != FunctionDefinition::Visibility::External)
-		m_errorReporter.typeError(fallback->location(), "Fallback function must be defined as \"external\".");
 }
 
 void ContractLevelChecker::checkExternalTypeClashes(ContractDefinition const& _contract)